display video on media carousel;
