+++ src/components/Programs/ProgramPage.vue
@@ -4,7 +4,11 @@
     <div v-if="program" class="pageTitle">
       <h1 class="pageTitleText">{{ program.name }}</h1>
     </div>
-    <Media v-if="program" :gallery="program.gallery" />
+    <Media
+      v-if="program"
+      :youtubeUrl="program.video_url"
+      :gallery="program.gallery"
+    />
     <div
       v-if="program"
       class="pageRow pageBody programText"
